VmCpu: Move external interrupt management into its own module
The external interrupt bits are pretty self-contained. Hoist them into
their own module.
